split sys_math, sys_math3d, z_lib from global.h
