+++ include/libultra_gu_math.h
@@ -1,0 +1,11 @@
+#ifndef Z_LIBULTRA_GU_MATH_H
+#define Z_LIBULTRA_GU_MATH_H
+#include "ultra64.h"
+
+f32 sinf(f32 angle);
+s16 sins(u16 angle);
+
+f32 cosf(f32 angle);
+s16 coss(u16 angle);
+
+#endif
