+++ include/regs.h
@@ -353,4 +353,15 @@ typedef enum {
 #define R_VI_CUR_ADDI_SCAN_LINES   HREG(83)
 #define R_VI_CUR_Y_SCALE_MODE      HREG(84)
 
+typedef struct {
+    /* 0x00 */ s32  regPage; // 0: no page selected (reg editor is not active); 1: first page; `REG_PAGES`: last page
+    /* 0x04 */ s32  regGroup; // Indexed from 0 to `REG_GROUPS`-1. Each group has its own character to identify it.
+    /* 0x08 */ s32  regCur; // Selected reg, indexed from 0 as the page start
+    /* 0x0C */ s32  dPadInputPrev;
+    /* 0x10 */ s32  inputRepeatTimer;
+    /* 0x14 */ s16  data[REG_GROUPS * REGS_PER_GROUP]; // Accessed through *REG macros, see regs.h
+} RegEditor; // size = 0x15D4
+
+extern RegEditor* gRegEditor;
+
 #endif
