+++ include/sys_math.h
@@ -1,0 +1,13 @@
+#ifndef SYS_MATH_H
+#define SYS_MATH_H
+#include "ultra64.h"
+
+f32 Math_FactorialF(f32 n);
+f32 Math_Factorial(s32 n);
+f32 Math_PowF(f32 base, s32 exp);
+f32 Math_SinF(f32 angle);
+f32 Math_CosF(f32 angle);
+
+#include "sys_math_atan.h"
+
+#endif
