+++ include/sys_math_atan.h
@@ -1,0 +1,8 @@
+#ifndef SYS_MATH_ATAN_H
+#define SYS_MATH_ATAN_H
+#include "ultra64.h"
+
+s16 Math_Atan2S(f32 x, f32 y);
+f32 Math_Atan2F(f32 x, f32 y);
+
+#endif
