+++ include/z64.h
@@ -59,6 +59,13 @@
 #include "jpeg.h"
 #include "prerender.h"
 
+// new headers
+#include "rand.h"
+#include "sys_math.h"
+#include "sys_math3d.h"
+#include "z_lib.h"
+#include "libultra_gu_math.h"
+
 #define SCREEN_WIDTH  320
 #define SCREEN_HEIGHT 240
 
@@ -98,15 +105,6 @@
 
 // NOTE: Once we start supporting other builds, this can be changed with an ifdef
 #define REGION_NATIVE REGION_EU
-
-typedef struct {
-    /* 0x00 */ s32  regPage; // 0: no page selected (reg editor is not active); 1: first page; `REG_PAGES`: last page
-    /* 0x04 */ s32  regGroup; // Indexed from 0 to `REG_GROUPS`-1. Each group has its own character to identify it.
-    /* 0x08 */ s32  regCur; // Selected reg, indexed from 0 as the page start
-    /* 0x0C */ s32  dPadInputPrev;
-    /* 0x10 */ s32  inputRepeatTimer;
-    /* 0x14 */ s16  data[REG_GROUPS * REGS_PER_GROUP]; // Accessed through *REG macros, see regs.h
-} RegEditor; // size = 0x15D4
 
 typedef struct {
     /* 0x00 */ u8   seqId;
