+++ include/z_lib.h
@@ -1,0 +1,6 @@
+#ifndef ICHAIN_H
+#define ICHAIN_H
+
+#include "ichain.h"
+
+#endif
