+++ src/code/rand.c
@@ -42,6 +42,7 @@
  * @note Original name: qrand.c
  */
 #include "ultra64.h"
+#include "rand.h"
 
 #define RAND_MULTIPLIER 1664525
 #define RAND_INCREMENT 1013904223
