+++ src/code/sys_math.c
@@ -1,4 +1,5 @@
-#include "global.h"
+#include "libultra_gu_math.h"
+#include "sys_math.h"
 
 f32 sFactorialTbl[] = { 1.0f,    1.0f,     2.0f,      6.0f,       24.0f,       120.0f,      720.0f,
                         5040.0f, 40320.0f, 362880.0f, 3628800.0f, 39916800.0f, 479001600.0f };
