+++ src/code/sys_math3d_draw.c
@@ -1,0 +1,10 @@
+#include "global.h"
+#include "terminal.h"
+
+#if OOT_DEBUG
+void Math3D_DrawSphere(PlayState* play, Sphere16* sph) {
+}
+
+void Math3D_DrawCylinder(PlayState* play, Cylinder16* cyl) {
+}
+#endif
