+++ src/code/z_kankyo.c
@@ -215,7 +215,7 @@ s16 sSunDepthTestY;
 // These variables could be moved farther down in the file to reduce the amount
 // of block number padding here, but currently this causes BSS ordering issues
 // for debug.
-#pragma increment_block_number 230
+#pragma increment_block_number 227
 
 LightNode* sNGameOverLightNode;
 LightInfo sNGameOverLightInfo;
