+++ src/code/z_lib.c
@@ -1,4 +1,18 @@
-#include "global.h"
+#include "ultra64.h"
+#include "z_intrinsic.h"
+#include "libultra_gu_math.h"
+#include "ichain.h" //#include "z_lib.h"
+#include "regs.h"
+#include "z_lib.h"
+#include "macros.h"
+#include "sys_math.h"
+#include "rand.h"
+#include "sfx.h"
+
+//! TODO: These should be moved into a header
+extern Vec3f gSfxDefaultPos;
+extern f32 gSfxDefaultFreqAndVolScale;
+extern s8 gSfxDefaultReverb;
 
 /**
  * memset: sets `len` bytes to `val` starting at address `dest`.
@@ -334,7 +348,7 @@ void (*sInitChainHandlers[])(u8* ptr, InitChainEntry* ichain) = {
     IChain_Apply_Vec3f, IChain_Apply_Vec3fdiv1000, IChain_Apply_Vec3s,
 };
 
-void Actor_ProcessInitChain(Actor* actor, InitChainEntry* ichain) {
+void Actor_ProcessInitChain(void* actor, InitChainEntry* ichain) {
     do {
         sInitChainHandlers[ichain->type]((u8*)actor, ichain);
     } while ((ichain++)->cont);
